add emoji config test to Channel
